+++ src_docs/source/releases.rst
@@ -3,6 +3,24 @@
 Releases 
 ########
 
+v0.7.0
+------
+
+Development version (dev branch)
+
+* SSL support for HTTP targets (no validation of CA)
+* SSL support for MQTT targets (no validation of CA)
+* SSL support for OTA (no validation of CA)
+* Breaking change: To simplify the internal structure the 
+  temp sensor adjustment is now stored in C. So if you have 
+  enabled this function using F you will need to go into 
+  the configuration and update the factor again.
+* Added error handling for calibration page.
+
+TODO: 
+Update docs, MQTT ssl is enabled using :8883 at end, http targets enables using prefix https://
+Note! Brewfather don't support SSL.
+
 v0.6.0
 ------